Mapping and visualization
routes.csv is contains the spatial and ecoregion information about the
routes. Downloaded directly from the BBS website.

The R script is intended to be a launching point for another script in
which we actually analyze the spatial data. It contains useful functions
and tricks. BBS Mapping and visualization is a repository for
visualization tricks. Currently it contains a few demos on how to use
some functions from the "maps" package, it contains a function to
extract route locations given the country, state and route #, and it
contains a quick demo of a useful function in the fields package on how
to calculate the distance between two points.